Read nextState in enable logic to avoid crude hack
SimpleValueParser has this really ugly hack, which is wholly unneccessary in
hindshight.  The hack is to allow us to react to the delay between the
current state and the read character.

This solution, in turn, does present problems: it may increase
the propogation delay of this path, possibly lowering our
peak clock speed.  That is a later bridge.

And with that, we set our record for longest commit message to commit
size ratio in this project.  Maybe I should make some of this a comment...